fixed single command exit shell
